requirement.txt
flask==2.2.2
easyocr==1.4.1
supabase==0.0.4
twilio==7.14.0
serial==3.5
requests==2.28.1
python-dotenv==0.20.0
